refactor: extract data from Board component
