+++ front/src/modules/ui/components/board/Board.tsx
@@ -16,43 +16,26 @@ const StyledBoard = styled.div`
   height: 100%;
 `;
 
-type ItemKey = `item-${number}`;
-interface Item {
+export type ItemKey = `item-${number}`;
+export interface Item {
   id: string;
   content: string;
 }
-interface Items {
+export interface Items {
   [key: string]: Item;
 }
-interface Column {
+export interface Column {
   id: string;
   title: string;
   itemKeys: ItemKey[];
 }
 
-const items: Items = {
-  'item-1': { id: 'item-1', content: 'Item 1' },
-  'item-2': { id: 'item-2', content: 'Item 2' },
-  'item-3': { id: 'item-3', content: 'Item 3' },
-  'item-4': { id: 'item-4', content: 'Item 4' },
-  'item-5': { id: 'item-5', content: 'Item 5' },
-  'item-6': { id: 'item-6', content: 'Item 6' },
-} satisfies Record<ItemKey, { id: ItemKey; content: string }>;
+type BoardProps = {
+  initialBoard: Column[];
+  items: Items;
+};
 
-const initialBoard = [
-  {
-    id: 'column-1',
-    title: 'Column 1',
-    itemKeys: ['item-1', 'item-2', 'item-3', 'item-4'],
-  },
-  {
-    id: 'column-2',
-    title: 'Column 2',
-    itemKeys: ['item-5', 'item-6'],
-  },
-] satisfies Column[];
-
-export const Board = () => {
+export const Board = ({ initialBoard, items }: BoardProps) => {
   const [board, setBoard] = useState<Column[]>(initialBoard);
 
   const onDragEnd: OnDragEndResponder = useCallback(
